+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 = 14.13.4 - 2025-**-** =
 - **Enhancement:** Improved privacy helper text.
+- **Enhancement:** Restricted migration process functionality to only run within the plugin's admin pages.
 
 = 14.13.3 - 2025-04-23 =
 - **New:** Added an option to enable or disable the Privacy Audit section in email reports.